Make more of a distinction between featured sites and other sites
Experimenting with a light blue background for the featured sites,
to make them stand out. Had some feedback that the sorting order
was confusing. I like the idea of keeping prominent sites on top
and alphabetizing the rest, so hopefully this will help as a visual
cue of where the sorting order changes.